#5881. libjpeg package updated from 9c with local modifications to 9d from conan-center.
The local modifications exposed jpegint.h with internals, which is used in ROQ tool.
However, the new recipe from conan-center exposes this header automatically, so we don't need to hack anything.

Had to do a few fixes in the code:
renderer/Image_files.cpp
  jpeg_mem_dest now outputs size into size_t instead of unsigned long
sys/msvc/properties/_DoomDLL.props
  define LIBJPEG_STATIC macro, otherwise compiler tries to use dllimport-ed symbols


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@9867 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ renderer/Image_files.cpp
@@ -680,7 +680,7 @@ bool idImageWriter::WriteJPG(int quality) {
 	 * VERY IMPORTANT: use "b" option to fopen() if you are on a machine that
 	 * requires it in order to write binary files.
 	 */
-	unsigned long outputSize;
+	size_t outputSize;
 	jpeg_mem_dest(&cinfo, &outputBuffer, &outputSize);
 
 	/* Step 3: set parameters for compression */
